🩹 Added conversion from int to float

+++ src/game_engine/systems/player_control_system.cpp
@@ -104,20 +104,20 @@ void sys::player_control_system(EngineContext& ctx, ecs::SparseArray<cpnt::Trans
                         if (sprite->frame <= k_animation_frame_limit && sprite->source_rect.x != k_sprite_up_right) {
                             sprite->frame++;
                             sprite->source_rect.x = k_sprite_up_anim;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                         } else {
                             sprite->source_rect.x = k_sprite_up_right;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                             sprite->frame = 0;
                         }
                     } else if (input.move_y > 0.0f) {
                         if (sprite->frame <= k_animation_frame_limit && sprite->source_rect.x != k_sprite_down_left) {
                             sprite->frame++;
                             sprite->source_rect.x = k_sprite_down_anim;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                         } else {
                             sprite->source_rect.x = k_sprite_down_left;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                             sprite->frame = 0;
                         }
                     } else {
@@ -125,16 +125,16 @@ void sys::player_control_system(EngineContext& ctx, ecs::SparseArray<cpnt::Trans
                                                                          sprite->source_rect.x == k_sprite_down_anim)) {
                             sprite->frame++;
                             sprite->source_rect.x = k_sprite_down_anim;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                         } else if (sprite->frame <= k_animation_frame_limit &&
                                    (sprite->source_rect.x == k_sprite_up_right ||
                                     sprite->source_rect.x == k_sprite_up_anim)) {
                             sprite->frame++;
                             sprite->source_rect.x = k_sprite_up_anim;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                         } else {
                             sprite->source_rect.x = k_sprite_neutral;
-                            sprite->source_rect.y = k_sprite_y_add * k_player_id;
+                            sprite->source_rect.y = k_sprite_y_add * (float)k_player_id;
                             sprite->frame = 0;
                         }
                     }
